añadido estado de tienda física y la manera de ir hasta él, pero no funcionan las colisiones de este nuevo estado ni se puede volver atrás todavía.

+++ Proyecto/RedBrickSky/RedBrickSky/GameManager.h
@@ -62,8 +62,8 @@ private:
 
 	static GameManager* s_pInstance;
 
-	float playerHP = 1000;
-	float playerMaxHP = 1000;
+	int playerHP = 1000;
+	int playerMaxHP = 1000;
 
 public:
 	~GameManager() {};
@@ -100,7 +100,7 @@ public:
 	int askAboutLevel() { return level; }
 	void setLevel(int l) { level = l; }
 
-	float getHealth() { return playerHP; };
-	float getMaxHealth() { return playerMaxHP; };
+	int getHealth() { return playerHP; };
+	int getMaxHealth() { return playerMaxHP; };
 	void setHealth(int h) { playerHP = h; };
 };
